--- data.json
@@ -1,112 +1,0 @@
-{
-  "instructors": [
-    {
-      "id": 1,
-      "avatar_url": "https://images.unsplash.com/photo-1438761681033-6461ffad8d80?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80",
-      "name": "Larissa Melo",
-      "gender": "F",
-      "birth": 593308800000,
-      "services": [
-        "Crossfit",
-        " Funcional"
-      ],
-      "created_at": 1593443790907
-    },
-    {
-      "id": 4,
-      "avatar_url": "https://images.unsplash.com/photo-1494790108377-be9c29b29330?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1868&q=80",
-      "name": "Joana",
-      "gender": "F",
-      "birth": 851040000000,
-      "services": [
-        "Musculação",
-        " Ginástica"
-      ],
-      "created_at": 1593469121501
-    },
-    {
-      "id": 5,
-      "avatar_url": "https://images.unsplash.com/photo-1506277886164-e25aa3f4ef7f?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=675&q=80",
-      "name": "Gabriel",
-      "gender": "M",
-      "birth": 537580800000,
-      "services": [
-        "Cross Training",
-        " Pilates",
-        " Funcional"
-      ],
-      "created_at": 1593618364334
-    },
-    {
-      "id": 6,
-      "avatar_url": "https://images.unsplash.com/photo-1488426862026-3ee34a7d66df?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=634&q=80",
-      "name": "Gabriela",
-      "gender": "F",
-      "birth": 788745600000,
-      "services": [
-        "Fit Dance"
-      ],
-      "created_at": 1593618564240
-    },
-    {
-      "id": 7,
-      "avatar_url": "https://images.unsplash.com/photo-1586798658502-0dc367cb2776?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=634&q=80",
-      "name": "Tiago Ventura",
-      "gender": "M",
-      "birth": 853718400000,
-      "services": [
-        "Natação",
-        " Judô",
-        " Basquete"
-      ],
-      "created_at": 1593799659195
-    },
-    {
-      "id": 8,
-      "avatar_url": "https://images.unsplash.com/photo-1463453091185-61582044d556?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80",
-      "name": "Clarice Lispector",
-      "gender": "F",
-      "birth": 706924800000,
-      "services": [
-        "Crossfit",
-        " Natação"
-      ],
-      "created_at": 1593884614719
-    }
-  ],
-  "members": [
-    {
-      "id": 2,
-      "avatar_url": "https://images.unsplash.com/photo-1438761681033-6461ffad8d80?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80",
-      "name": "Jeana",
-      "email": "explaing@aulas.com",
-      "gender": "F",
-      "birth": 524448000000,
-      "blood": "B+",
-      "height": "181",
-      "weight": "65"
-    },
-    {
-      "id": 3,
-      "avatar_url": "https://images.unsplash.com/photo-1591877499903-a8ca88dd7f60?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=500&q=60",
-      "name": "Jonas",
-      "email": "explaing@aulas.com",
-      "gender": "M",
-      "birth": 930096000000,
-      "blood": "A+",
-      "height": "181",
-      "weight": "88"
-    },
-    {
-      "id": 4,
-      "avatar_url": "https://images.unsplash.com/photo-1438761681033-6461ffad8d80?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80",
-      "name": "Janice",
-      "email": "janice@gmail.com",
-      "gender": "F",
-      "birth": 712022400000,
-      "blood": "AB-",
-      "height": "165",
-      "weight": "55"
-    }
-  ]
-}
