+++ package.json
@@ -13,7 +13,7 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "npm-run-all -p nodemon browsersync",
     "nodemon": "nodemon src/server.js",
-    "browsersync": "browser-sync start --proxy http://localhost:3000 --files 'public,views'"
+    "browsersync": "browser-sync start --proxy http://localhost:3000 --files 'public,src/app/views'"
   },
   "keywords": [],
   "author": "",
