+++ src/app/controllers/instructors.js
@@ -1,6 +1,4 @@
-const fs = require('fs');
-const data = require('../../data.json');
-const { age, date } = require('../utils');
+const { age, date } = require('../lib/utils');
 Intl = require('intl');
