+++ src/app/views/instructors/index.njk
