+++ src/app/views/instructors/show.njk
