+++ src/routes.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const routes = express.Router();
-const instructors = require('./controllers/instructors');
-const members = require('./controllers/members');
+const instructors = require('./app/controllers/instructors');
+const members = require('./app/controllers/members');
 
 routes.get('/', (req,res) => {
   return res.redirect('/instructors');
