+++ src/server.js
@@ -9,7 +9,7 @@ const port = process.env.PORT || 3000;
 
 
 server.set('view engine', 'njk');
-nunjucks.configure("views", {
+nunjucks.configure("src/app/views", {
   express: server,
   autoescape: false,
   noCache: true,
